refactor by splitting out confirmation dialog
